fix: update stock balances when production outputs change
Adjust stock balance quantities when production outputs are created, updated, or relocated. This ensures inventory accuracy by synchronizing production output changes with stock balances.

- Change stock_balances.quantity from decimal to integer for consistency with production output quantities
- Add stock balance adjustments in storeBatchAndOutputs for new outputs
- Add stock balance adjustments in updateBatchAndOutputs for quantity changes and location moves
- Use lockForUpdate to prevent race conditions during concurrent updates

+++ database/migrations/2026_02_21_000011_create_stock_balances_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->uuid('owner_id');
             $table->uuid('item_id');
             $table->uuid('location_id');
-            $table->decimal('quantity', 14, 2);
+            $table->integer('quantity');
 
             $table->primary(['owner_id', 'item_id', 'location_id']);
